+++ internal/biz/common.go
@@ -1,4 +1,4 @@
-package data
+package biz
 
 import (
 	"context"
